Move some solver stuff to middle
